feat(image resizer): raw image bytes
The image resizing logic has been moved to a separate crate (image_worker) to make dev builds faster. Also, the images are no longer converted to a PNG and instead rendered as to a HTML <canvas> as raw pixel data (bytes).

Images are no longer returned as a PNG but rather raw pixel bytes.